+++ tanaguru-engine/tanaguru-audit-runner/src/main/java/com/tanaguru/service/impl/runner/AuditRunnerServiceAsyncStandaloneImpl.java
@@ -49,7 +49,6 @@ public class AuditRunnerServiceAsyncStandaloneImpl extends AbstractAuditRunnerSe
     public AuditRunnerServiceAsyncStandaloneImpl(
             PageRepository pageRepository,
             AuditRepository auditRepository,
-            AuditRunnerFactory auditRunnerFactory,
             AuditService auditService,
             PageContentRepository pageContentRepository,
             TestResultRepository testResultRepository,
@@ -60,7 +59,10 @@ public class AuditRunnerServiceAsyncStandaloneImpl extends AbstractAuditRunnerSe
             MailService mailService,
             MessageService messageService,
             ActRepository actRepository,
-            ContractUserRepository contractUserRepository) {
+            ContractUserRepository contractUserRepository,
+            ProjectUserRepository projectUserRepository,
+            AuditRunnerFactory auditRunnerFactory
+    ) {
         super(pageRepository,
                 auditRepository,
                 auditService,
@@ -73,7 +75,7 @@ public class AuditRunnerServiceAsyncStandaloneImpl extends AbstractAuditRunnerSe
                 mailService,
                 messageService,
                 actRepository,
-                contractUserRepository);
+                contractUserRepository, projectUserRepository);
         this.auditRunnerFactory = auditRunnerFactory;
     }
 
@@ -140,22 +142,38 @@ public class AuditRunnerServiceAsyncStandaloneImpl extends AbstractAuditRunnerSe
         }
     }
 
+    @Override
+    public void stopAudit(Audit audit){
+        for (AuditRunner runner : concurrentAuditRunnerMap.keySet()) {
+            if(runner.getAudit().getId() == audit.getId()){
+                LOGGER.warn("[Audit {}] Interrupting audit", runner.getAudit().getId());
+                auditService.log(runner.getAudit(), EAuditLogLevel.WARNING, "Audit Interrupted by server");
+                runner.interrupt();
+            }
+        }
+    }
+
     /**
      * Hooks kill event, cleans audit that will not be launched
      */
     @PreDestroy
-    private void cleanRunningAudits() {
+    private void cleanRunningAudits() throws InterruptedException {
         synchronized (waitingRequests) {
             for (Audit audit : waitingRequests) {
-                auditService.log(audit, EAuditLogLevel.ERROR, "Audit Interrupted by server");
+                auditService.log(audit, EAuditLogLevel.WARNING, "Audit Interrupted by server");
+                waitingRequests.remove(audit);
             }
         }
 
         synchronized (concurrentAuditRunnerMap) {
             for (AuditRunner runner : concurrentAuditRunnerMap.keySet()) {
                 LOGGER.warn("[Audit {}] Interrupting audit", runner.getAudit().getId());
-                auditService.log(runner.getAudit(), EAuditLogLevel.ERROR, "Audit Interrupted by server");
-                concurrentAuditRunnerMap.get(runner).interrupt();
+                auditService.log(runner.getAudit(), EAuditLogLevel.WARNING, "Audit Interrupted by server");
+                runner.interrupt();
+            }
+
+            for (AuditRunner runner : concurrentAuditRunnerMap.keySet()) {
+                concurrentAuditRunnerMap.get(runner).join();
             }
         }
     }