+++ CHANGELOG.md
@@ -22,6 +22,8 @@
 - Translation of the mail sent when the user is blocked
 - Endpoint : audit logs filtered by date and/or level
 - Pagination for contracts, users, references, projects
+- Fix mail at the end of audit (was sending to all contract users)
+- Endpoint stop audit for async audit (other profiles will throw a bad request exception)
 - Add configuration key to enable or not robots.txt respect (auditrunner.crawler.follow-robots)
 - Optimize TestResult getAllTestResultByReference endpoint
 - Add a parameter on createUser endpoint to create a contract